`controller`:split_into->{[`controller`|`controller_state`]}.further_refactor
