+++ sources/metil_menus/menu.c
@@ -1,6 +1,6 @@
 #include <metil_menus/menu.h>
 
-#include <metil_input/controller.h>
+#include <metil_input/controller_state.h>
 #include <metil_input/keycodes.h>
 #include <metil_input/map.h>
 #include <metil_menus/menu_item.h>
@@ -42,7 +42,7 @@ void metil_menu_poll_input(
     metil_input_map_keydown[
       metil_keycode_space
     ] == 1 ||
-    metil_controller_state.button_cross > 0.0f
+    metil_controller_state.cross > 0.0f
   ) {
     metil_menu_select(
       menu
@@ -67,7 +67,7 @@ void metil_menu_poll_input(
     metil_input_map_keydown[
       metil_keycode_up_arrow
     ] == 1 || 
-    metil_controller_state.button_directional_up > 0.0f ||
+    metil_controller_state.directional_up > 0.0f ||
     metil_controller_state.left_stick.y > 0.1f
   ) {
     had_input = 1;
@@ -79,7 +79,7 @@ void metil_menu_poll_input(
     metil_input_map_keydown[
       metil_keycode_down_arrow
     ] == 1 || 
-    metil_controller_state.button_directional_down > 0.0f ||
+    metil_controller_state.directional_down > 0.0f ||
     metil_controller_state.left_stick.y < -0.1f
   ) {
     had_input = 1;
